@cmsbuild 2015-07-20
A new Pull Request was created by @dmitrijus (Dmitrijus) for CMSSW_7_4_X.

Implement multi-processing in fastHadd to merge DQM histograms faster.

It involves the following packages:

DQMServices/Components


@cmsbuild, @danduggan, @deguio can you please review it and eventually sign? Thanks.
@barvic this is something you requested to watch as well.
You can sign-off by replying to this message having '+1' in the first line of your reply.
You can reject by replying  to this message having '-1' in the first line of your reply.
If you are a L2 or a release manager you can ask for tests by saying 'please test' in the first line of a comment.
@degano you are the release manager for this.
You can merge this pull request by typing 'merge' in the first line of your comment.