@deguio 2015-07-27
+1
I validated the content and I confirm 100% success in the following cases:
- standard fastHadd VS new version with 1 node
- standard fastHadd VS new version with 7 nodes

the performance are also confirmed.

@dmitrijus 
please submit to 75 and 76 as well.
thanks,
F.