@deguio 2015-07-27
hello @davidlange6 
can we have this one merged? this is used in the online world and to propagate it to the prod system we need to submit a PR to cmsdist as well.
ideally we would like to prepare the RPM and ask for its installation tomorrow.
thanks,
F.