@cmsbuild 2015-07-27
+1
Tested at: b5439df5f22b94107fda9472fb6007b3dd9b6a79
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10280/6642/summary.html