@cmsbuild 2015-07-27
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10280/6642/summary.html