@smuzaffar 2017-09-04
@dmitrijus , this PR (though very old) had removed  DQMServices/Components/test/fastParallelHadd.py which is needed here https://github.com/cms-sw/cmssw/blob/master/DQMServices/Components/test/run_fastHadd_tests.sh#L115 

Can you please provide a fix for the and also @Dr15Jones noticed that run_fastHadd_tests.sh runs ```rm -fr *.root``` which deletes root files needed by other unittests. Can you please make sure that run_fastHadd_tests.sh create a sub-directory e.g.  runFastHadd and it creates/deletes file only under this sub-dir?